+++ k8s-eks-addons.tf
@@ -3,6 +3,7 @@ module "k8s_eks_addons" {
 
   ingress_nginx_config      = merge(var.ingress_nginx_config, { subnets_ids = local.public_subnets })
   cluster_autoscaler_config = var.cluster_autoscaler_config
+  coredns_config            = var.coredns_config
 
   addon_context = {
     aws_caller_identity_account_id = data.aws_caller_identity.current.account_id