+++ modules/k8s_eks_addons/coredns.tf
@@ -18,5 +18,4 @@ resource "aws_eks_addon" "coredns" {
   resolve_conflicts_on_update = "OVERWRITE"
   configuration_values        = var.coredns_config.configuration_values
   tags                        = var.addon_context.tags
-
 }